Move AutoPadType to common.h
