+++ onnxruntime/core/providers/acl/nn/batch_norm.cc
@@ -5,7 +5,7 @@
 #include "core/common/common.h"
 #include "core/common/exceptions.h"
 #include "core/framework/op_kernel.h"
-#include "core/providers/cpu/nn/autopad_type.h"
+#include "core/providers/common.h"
 #include "core/framework/tensor.h"
 #include "core/util/math_cpuonly.h"
