+++ onnxruntime/core/providers/cpu/nn/roi_pool.h
@@ -5,7 +5,7 @@
 
 #include "core/common/common.h"
 #include "core/framework/op_kernel.h"
-#include "core/providers/cpu/nn/autopad_type.h"
+#include "core/providers/common.h"
 
 namespace onnxruntime {
