+++ onnxruntime/core/providers/cpu/nn/unpool.h
@@ -6,7 +6,7 @@
 #include <cmath>
 #include "core/common/common.h"
 #include "core/framework/op_kernel.h"
-#include "core/providers/cpu/nn/autopad_type.h"
+#include "core/providers/common.h"
 
 namespace onnxruntime {
